Refactor activity commands
