+++ cli/activity.go
@@ -25,6 +25,9 @@
 package cli
 
 import (
+	"fmt"
+
+	"github.com/temporalio/tctl/pkg/color"
 	"github.com/urfave/cli/v2"
 )
 
@@ -54,6 +57,11 @@ func newActivityCommands() []*cli.Command {
 				&cli.StringFlag{
 					Name:  FlagIdentity,
 					Usage: "Identity of the operator",
+				},
+				&cli.StringFlag{
+					Name:  color.FlagColor,
+					Usage: fmt.Sprintf("when to use color: %v, %v, %v.", color.Auto, color.Always, color.Never),
+					Value: string(color.Auto),
 				},
 			},
 			Action: func(c *cli.Context) error {
@@ -88,6 +96,11 @@ func newActivityCommands() []*cli.Command {
 				&cli.StringFlag{
 					Name:  FlagIdentity,
 					Usage: "Identity of the operator",
+				},
+				&cli.StringFlag{
+					Name:  color.FlagColor,
+					Usage: fmt.Sprintf("when to use color: %v, %v, %v.", color.Auto, color.Always, color.Never),
+					Value: string(color.Auto),
 				},
 			},
 			Action: func(c *cli.Context) error {
