+++ cli/workflowCommands.go
@@ -1362,74 +1362,6 @@ func getLastContinueAsNewID(ctx context.Context, namespace, wid, rid string, fro
 	return
 }
 
-// CompleteActivity completes an activity
-func CompleteActivity(c *cli.Context) {
-	namespace := getRequiredGlobalOption(c, FlagNamespace)
-	wid := getRequiredOption(c, FlagWorkflowID)
-	rid := getRequiredOption(c, FlagRunID)
-	activityID := getRequiredOption(c, FlagActivityID)
-	if len(activityID) == 0 {
-		ErrorAndExit("Invalid activityId", fmt.Errorf("activityId cannot be empty"))
-	}
-	result := getRequiredOption(c, FlagResult)
-	identity := getRequiredOption(c, FlagIdentity)
-	ctx, cancel := newContext(c)
-	defer cancel()
-
-	frontendClient := cFactory.FrontendClient(c)
-	_, err := frontendClient.RespondActivityTaskCompletedById(ctx, &workflowservice.RespondActivityTaskCompletedByIdRequest{
-		Namespace:  namespace,
-		WorkflowId: wid,
-		RunId:      rid,
-		ActivityId: activityID,
-		Result:     payloads.EncodeString(result),
-		Identity:   identity,
-	})
-	if err != nil {
-		ErrorAndExit("Completing activity failed", err)
-	} else {
-		fmt.Println("Complete activity successfully.")
-	}
-}
-
-// FailActivity fails an activity
-func FailActivity(c *cli.Context) {
-	namespace := getRequiredGlobalOption(c, FlagNamespace)
-	wid := getRequiredOption(c, FlagWorkflowID)
-	rid := getRequiredOption(c, FlagRunID)
-	activityID := getRequiredOption(c, FlagActivityID)
-	if len(activityID) == 0 {
-		ErrorAndExit("Invalid activityId", fmt.Errorf("activityId cannot be empty"))
-	}
-	reason := getRequiredOption(c, FlagReason)
-	detail := getRequiredOption(c, FlagDetail)
-	identity := getRequiredOption(c, FlagIdentity)
-	ctx, cancel := newContext(c)
-	defer cancel()
-
-	frontendClient := cFactory.FrontendClient(c)
-	_, err := frontendClient.RespondActivityTaskFailedById(ctx, &workflowservice.RespondActivityTaskFailedByIdRequest{
-		Namespace:  namespace,
-		WorkflowId: wid,
-		RunId:      rid,
-		ActivityId: activityID,
-		Failure: &failurepb.Failure{
-			Message: reason,
-			Source:  "CLI",
-			FailureInfo: &failurepb.Failure_ApplicationFailureInfo{ApplicationFailureInfo: &failurepb.ApplicationFailureInfo{
-				NonRetryable: true,
-				Details:      payloads.EncodeString(detail),
-			}},
-		},
-		Identity: identity,
-	})
-	if err != nil {
-		ErrorAndExit("Failing activity failed", err)
-	} else {
-		fmt.Println("Fail activity successfully.")
-	}
-}
-
 func getWorkflowParams(c *cli.Context) (string, string) {
 	var wid, rid string
